added authentication

+++ controllers/feed.js
@@ -2,6 +2,7 @@ const { validationResult } = require('express-validator/check')
 const path = require("path")
 const fs = require("fs")
 const Post = require("../models/post")
+const User = require("../models/user")
 
 exports.getPost = (req, res, next) => {
   const { postId } = req.params
@@ -54,6 +55,8 @@ exports.createPost = (req, res, next) => {
   const errors = validationResult(req)
   const title = req.body.title;
   const content = req.body.content;
+  let posts;
+  let creator;
   // Create post in db
   if (!errors.isEmpty()) {
     const error = new Error("Validation Failed")
@@ -73,16 +76,26 @@ exports.createPost = (req, res, next) => {
     title: title,
     content: content,
     imageUrl: req.file.path,
-    creator: {
-      name: 'Maximilian'
-    },
+    creator: req.userId
   })
 
   post.save().then((data) => {
+    posts = data
+    return User.findById(req.userId)
+  }).then(user => {
+    creator = user
+    user.posts.push(posts)
+    return user.save()
+  }).then(result => {
     res.status(201).json({
-      message: 'Post created successfully!',
-      post: data
-    });
+      message: "Post created!",
+      post: post,
+      creator: {
+        _id: creator._id,
+        name: creator.name
+      }
+    })
+
   }).catch((err) => {
     // db error status code must be 500
     if (!err.statusCode) {
@@ -130,6 +143,12 @@ exports.updatePosts = (req, res, next) => {
       // throw err will take to catch from then
     }
 
+    if (post.creator.toString() !== req.userId) {
+      const err = new Error("Unauthorized!")
+      err.statusCode = 403
+      throw err
+    }
+
     if (imageUrl !== post.imageUrl) {
       clearImage(post.imageUrl)
     }
@@ -164,10 +183,22 @@ exports.deletePost = (req, res, next) => {
 
       // throw err will take to catch from then
     }
+
+    if (post.creator.toString() !== req.userId) {
+      const err = new Error("Unauthorized!")
+      err.statusCode = 403
+      throw err
+    }
+
     // check the logged in user
     clearImage(post.imageUrl)
 
     return Post.findByIdAndRemove(postId)
+  }).then((result) => {
+    return User.findById(req.userId)
+  }).then((user) => {
+    user.posts.pull(postId)
+    return user.save()
   }).then((result) => {
     res.status(200).json({
       message: "Post deleted!",
